src/log-analysis/automated-tarball-ingestion/requirements.txt
PyYAML
# Elasticsearch 7.x
elasticsearch>=7.0.0,<8.0.0
